backend/requirements.txt
fastapi==0.124.0
uvicorn[standard]==0.38.0
coc.py==4.0.0
pydantic==2.12.5
httpx==0.28.1
sqlalchemy==2.0.44
python-dotenv==1.2.1
